bug: Add `mariadb` as required package to Dockerfile
Stage builds are failing with a `cffi` error. While I'm not able to
reproduce this locally, there are issues with the `alpine` docker base
no longer including the `libmysqlclient-dev` image.

Note: while running locally, I needed to specify the MOZSVC_SQLURI env
var. Not sure if the more clever interpolation broke recently, or was
always broken.

Issue: #187

+++ tokenserver/assignment/sqlnode/sql.py
@@ -362,7 +362,7 @@ class SQLNodeAssignment(object):
         migrate = self.get_migration_percent()
         if migrate:
             pick = ord(hashlib.sha1(email.encode()).digest()[0])
-            return pick < (256 * (migrate * .01))
+            return pick < int(256 * (migrate * .01))
         else:
             return False
